+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - improved error handling of lambda
+- fix race condition in `PipelineChannel`
 
 ## [0.0.38] - 2020-02-29